:zap:add httpclient

+++ openjob-server/openjob-server-admin/src/main/java/io/openjob/server/admin/vo/alert/ListAlertRuleVO.java
@@ -21,7 +21,7 @@ public class ListAlertRuleVO {
     private String name;
 
     @ApiModelProperty(value = "Name", required = true)
-    private List<Integer> namespaceAppIds;
+    private List<Long> namespaceAppIds;
 
     @ApiModelProperty(value = "Events")
     private List<String> events;
